[REFACTOR][TIRX] Use single-inheritance typed StmtExprMutator (#20365)
Replace the separate TIRx expression and statement mutator hierarchy
with a single typed dispatch chain through the shared ExprMutator.
Statement-only callers keep explicit expression boundaries; specialized
passes retain their ownership, geometry and type-inference
responsibilities.

Canonical variable and buffer replacements share the inherited
definition environment. Dtype normalization separates collection from
rewriting, and compute legalization uses one candidate map with
opaque-use filtering before mutation. Native and structural SeqStmt
mutation share the original structural algorithm and its in-place
sequence handling. Ordinary memory-access hooks retain their
pass-specific buffer policy and stored types.

Mutation planning uses the native visitor API. Statement simplification
preserves buffer definitions and delegates each primitive expression
directly to the analyzer. Buffer pointer hooks update the final result
type after parent mutation, preserving other call fields.

IKET cleanup runs after ordinary parent mutation, retaining only
empty-body and empty-branch removal.

+++ tests/python/relax/test_transform_legalize_ops_grad.py
@@ -187,12 +187,12 @@ def test_nll_loss_backward_no_batch():
                 T.writes(all_weights[()])
                 all_weights[()] = rxplaceholder_3[rxplaceholder_2[()]]
             with T.sblock("T_broadcast_to"):
-                vi = T.axis.spatial(1, T.int64(0))
+                vi = T.axis.spatial(T.int64(1), T.int64(0))
                 T.reads(rxplaceholder[()])
                 T.writes(T_broadcast_to[()])
                 T_broadcast_to[()] = rxplaceholder[()]
             with T.sblock("T_divide"):
-                vi = T.axis.spatial(1, T.int64(0))
+                vi = T.axis.spatial(T.int64(1), T.int64(0))
                 T.reads(T_broadcast_to[()], all_weights[()])
                 T.writes(T_divide[()])
                 T_divide[()] = T_broadcast_to[()] / all_weights[()]
